検索フィールド作成
