+++ src/app/components/editor/plugins/footer.tsx
@@ -3,6 +3,7 @@ import { Height } from '@mui/icons-material';
 import SendIcon from '@mui/icons-material/Send';
 import { Box, Button, IconButton, TextField } from '@mui/material';
 import { FC } from 'react';
+import { BaseTextField } from '../../ui/base-text-field';
 
 function downloadJSON(data: string, filename = 'data.json') {
     const blob = new Blob([data], { type: 'application/json' });
@@ -16,30 +17,6 @@ function downloadJSON(data: string, filename = 'data.json') {
     document.body.removeChild(a);
     URL.revokeObjectURL(url);
 }
-
-const textFieldStyle = {
-    width: 300,
-    '& .MuiInputBase-input': {
-        color: '#c3b0fa', // 入力文字の色
-    },
-    '& label': {
-        color: '#c3b0fa', // 通常時のラベル色
-    },
-    '& .MuiInput-underline:before': {
-        borderBottomColor: '#c3b0fa', // 通常時のボーダー色
-    },
-    '& .MuiInput-underline:hover:not(.Mui-disabled):before': {
-        borderBottomColor: '#c3b0fa', // ホバー時のボーダー色
-    },
-    '& .MuiOutlinedInput-root': {
-        '& fieldset': {
-            borderColor: '#c3b0fa', // 通常時のボーダー色(アウトライン)
-        },
-        '&:hover fieldset': {
-            borderColor: '#c3b0fa', // ホバー時のボーダー色(アウトライン)
-        },
-    },
-};
 
 export const FooterPlugin: FC = () => {
     const [editor] = useLexicalComposerContext();
@@ -65,12 +42,7 @@ export const FooterPlugin: FC = () => {
                 padding: '20px',
             }}
         >
-            <TextField
-                id='standard-basic'
-                label='Tags'
-                variant='standard'
-                sx={textFieldStyle}
-            />
+            <BaseTextField width={300} label='Tags' />
             <IconButton aria-label='delete' onClick={() => postTips()}>
                 <SendIcon sx={{ fontSize: 20, color: '#c3b0fa' }} />
             </IconButton>
